+++ global_trotter_backend/app.py
@@ -2,6 +2,7 @@ from flask import Flask
 from flask_jwt_extended import JWTManager
 
 from config import Config
+from routes.auth import auth_bp
 
 
 
@@ -10,6 +11,8 @@ def create_app():
     app.config.from_object(Config)
 
     JWTManager(app)
+
+    app.register_blueprint(auth_bp)